agents/fleet_mcp_client/requirements.txt
google-cloud-aiplatform[adk,agent_engines]==1.91.0
Flask==3.1.0
mcp[sse] >= 1.4.3
mcp[cli] >= 1.9.0
google-adk >= 1.0.0
python-dateutil==2.9.0.post0
humanize==4.12.3
nest_asyncio==1.6.0
asyncclick==8.1.8.0
a2a_common-0.1.0-py3-none-any.whl

